fix(daily_banking): 🐛 fix after solving merge conflicts

+++ lib/application/daily_banking/accounts/wires/sent_transfers/simplified/simplified_sent_transfers_controller.dart
@@ -81,11 +81,11 @@ class SimplifiedSentTransfersController extends StateNotifier<SimplifiedSentTran
     state = state.copyWith(endDate: endDate);
   }
 
-  void setAmountFrom(int? amountFrom) {
+  void setAmountFrom(double? amountFrom) {
     state = state.copyWith(amountFrom: amountFrom);
   }
 
-  void setAmountTo(int? amountTo) {
+  void setAmountTo(double? amountTo) {
     state = state.copyWith(amountTo: amountTo);
   }
 }
